requirements.txt
jira==3.5.1
gitpython==3.1.32
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
pygments==2.16.1
openai==1.3.5  # Optional: for AI-powered summaries
keyring==24.2.0  # For secure credential storage 